requirements.txt
pymupdf
python-docx
pandas
sentence-transformers
chromadb